+++ packages/web3-providers/src/TokenPrice/index.ts
@@ -1,5 +1,4 @@
 import type { CurrencyType } from '@masknet/plugin-infra'
-import type { CryptoPrice } from '@masknet/web3-shared-evm'
 import type { TokenPriceAPI } from '../types'
 
 const URL_BASE = 'https://api.coingecko.com/api/v3'
@@ -8,12 +7,12 @@ export class TokenPrice implements TokenPriceAPI.Provider {
     async getTokenPrices(platform: string, contractAddresses: string[], currency: CurrencyType) {
         const addressList = contractAddresses.join(',')
         const requestPath = `${URL_BASE}/simple/token_price/${platform}?contract_addresses=${addressList}&vs_currencies=${currency}`
-        const prices = await fetch(requestPath).then((r) => r.json() as Promise<CryptoPrice>)
+        const prices = await fetch(requestPath).then((r) => r.json() as Promise<TokenPriceAPI.CryptoPrice>)
         return prices
     }
     async getNativeTokenPrice(tokenIds: string[], currency: CurrencyType) {
         const requestPath = `${URL_BASE}/simple/price?ids=${tokenIds.join(',')}&vs_currencies=${currency}`
-        const prices = await fetch(requestPath).then((r) => r.json() as Promise<CryptoPrice>)
+        const prices = await fetch(requestPath).then((r) => r.json() as Promise<TokenPriceAPI.CryptoPrice>)
         return prices
     }
 }
